Fix up the website package
